requirements.txt
Django==5.0.6
numpy==2.0.0
Pillow==10.3.0
scikit_learn==1.5.0
